Teste unitário básico de validação de leituras do sonar. Correção de bugs no Sonar.

+++ Sauron/Sonar/CustomTypes.h
@@ -2,6 +2,6 @@
 
 namespace sauron
 {
-	typedef float reading_t;
-	typedef float pose_t;
+	typedef int reading_t;
+	typedef double pose_t;
 }
